[core] Tidy-up namespaces
Move all of the ORE related code into risk.
