+++ projects/ores.core/ores.core.hpp
