+++ projects/ores.core/risk/types/currency_config.hpp
@@ -1,6 +1,6 @@
 /* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * Copyright (C) 2024 Marco Craveiro <marco.craveiro@gmail.com>
+ * Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -17,16 +17,28 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#include <ostream>
-#include <rfl.hpp>
-#include <rfl/json.hpp>
-#include "ores.core/risk/currency_config.hpp"
+#ifndef ORES_CORE_RISK_TYPES_CURRENCY_CONFIG_HPP
+#define ORES_CORE_RISK_TYPES_CURRENCY_CONFIG_HPP
 
-namespace ores::core::risk {
+#if defined(_MSC_VER) && (_MSC_VER >= 1200)
+#pragma once
+#endif
 
-std::ostream& operator<<(std::ostream& s, const currency_config& v) {
-    rfl::json::write(v, s);
-    return(s);
+#include <iosfwd>
+#include <vector>
+#include "ores.core/risk/types/currency.hpp"
+
+namespace ores::core::risk::types {
+
+/**
+ * @brief Represents a set of ORE currencies.
+ */
+struct currency_config {
+    std::vector<currency> currencies;
+};
+
+std::ostream& operator<<(std::ostream& s, const currency_config& v);
+
 }
 
-}
+#endif
